+++ NEWS.md
@@ -1,10 +1,11 @@
 # Change log of the R package 'rgplates'
 
-# rgplates 0.6.1b - 2025-05-12
+# rgplates 0.6.1-3 - 2025-06-03
 
 ### Added
 
 - The `crs` argument of `mapedge` to project the created map edge immediately after creating it.
+- The `anchor` parameter is now available for the offline module of `reconstruct`.
 
 ### Fixed